+++ srunner/challenge/autoagents/autonomous_agent.py
@@ -1,0 +1,83 @@
+from enum import Enum
+from srunner.challenge.envs.sensor_interface import SensorInterface
+
+class Track(Enum):
+    """
+    This enum represents the different tracks of the CARLA AD challenge.
+    """
+    ALL_SENSORS = 1
+    CAMERAS = 2
+    ALL_SENSORS_HDMAP_WAYPOINTS = 3
+    SCENE_LAYOUT = 4
+
+class AutonomousAgent():
+    def __init__(self, path_to_conf_file):
+        self.track = Track.CAMERAS
+        #  current global plans to reach a destination
+        self._global_plan = None,
+
+        # this data structure will contain all sensor data
+        self.sensor_interface  = SensorInterface()
+
+        # agent's initialization
+        self.setup(path_to_conf_file)
+
+    def setup(self, path_to_conf_file):
+        """
+        Initialize everything needed by your agent and set the track attribute to the right type:
+            Track.ALL_SENSORS : LIDAR, cameras, GPS and speed sensor allowed
+            Track.CAMERAS : Only cameras and GPS allowed
+            Track.ALL_SENSORS_HDMAP_WAYPOINTS : All sensors and HD Map and waypoints allowed
+            Track.SCENE_LAYOUT : No sensors allowed, the agent receives a high-level representation of the scene.
+        """
+        pass
+
+    def sensors(self):
+        """
+        Define the sensor suite required by the agent
+
+        :return: a list containing the required sensors in the following format:
+
+        [
+            {'type': 'sensor.camera.rgb', 'x': 0.7, 'y': -0.4, 'z': 1.60, 'roll': 0.0, 'pitch': 0.0, 'yaw': 0.0,
+                      'width': 300, 'height': 200, 'fov': 100, 'id': 'Left'},
+
+            {'type': 'sensor.camera.rgb', 'x': 0.7, 'y': 0.4, 'z': 1.60, 'roll': 0.0, 'pitch': 0.0, 'yaw': 0.0,
+                      'width': 300, 'height': 200, 'fov': 100, 'id': 'Right'},
+
+            {'type': 'sensor.lidar.ray_cast', 'x': 0.7, 'y': 0.0, 'z': 1.60, 'yaw': 0.0, 'pitch': 0.0, 'roll': 0.0,
+             'id': 'LIDAR'}
+        ]
+
+        """
+        sensors = []
+
+        return sensors
+
+    def run_step(self):
+        """
+        Execute one step of navigation.
+        :return: control
+        """
+        pass
+
+    def destroy(self):
+        """
+        Destroy (clean-up) the agent
+        :return:
+        """
+        pass
+
+    def __call__(self):
+        input_data = self.sensor_interface.get_data()
+
+        control = self.run_step(input_data)
+        control.manual_gear_shift = False
+
+        return control
+
+    def all_sensors_ready(self):
+        return self.sensor_interface.all_sensors_ready()
+
+    def set_global_plan(self, global_plan):
+        self._global_plan = global_plan