+++ srunner/challenge/run_evaluator.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+python ${ROOT_SCENARIO_RUNNER}/srunner/challenge/challenge_evaluator.py  \
+--file \
+--scenario=group:ChallengeBasic \
+--agent=${ROOT_SCENARIO_RUNNER}/srunner/challenge/autoagents/HumanAgent.py \
+--route-visible