+++ srunner/scenariomanager/scenario_manager.py
@@ -18,9 +18,11 @@ import threading
 
 import py_trees
 
+import srunner
 from srunner.scenariomanager.carla_data_provider import CarlaDataProvider
 from srunner.scenariomanager.result_writer import ResultOutputProvider
 from srunner.scenariomanager.timer import GameTime, TimeOut
+from srunner.scenariomanager.traffic_events import TrafficEvent, TrafficEventType
 
 
 class Scenario(object):
@@ -43,21 +45,23 @@ class Scenario(object):
         self.test_criteria = criteria
         self.timeout = timeout
 
-        for criterion in self.test_criteria:
-            criterion.terminate_on_failure = terminate_on_failure
+        if not isinstance(self.test_criteria, py_trees.composites.Parallel):
+        # list of nodes
+            for criterion in self.test_criteria:
+                criterion.terminate_on_failure = terminate_on_failure
 
-        # Create py_tree for test criteria
-        self.criteria_tree = py_trees.composites.Parallel(name="Test Criteria")
-        self.criteria_tree.add_children(self.test_criteria)
-        self.criteria_tree.setup(timeout=1)
+            # Create py_tree for test criteria
+            self.criteria_tree = py_trees.composites.Parallel(name="Test Criteria")
+            self.criteria_tree.add_children(self.test_criteria)
+            self.criteria_tree.setup(timeout=1)
+        else:
+            self.criteria_tree = criteria
 
         # Create node for timeout
         self.timeout_node = TimeOut(self.timeout, name="TimeOut")
 
         # Create overall py_tree
-        self.scenario_tree = py_trees.composites.Parallel(
-            name,
-            policy=py_trees.common.ParallelPolicy.SUCCESS_ON_ONE)
+        self.scenario_tree = py_trees.composites.Parallel(name, policy=py_trees.common.ParallelPolicy.SUCCESS_ON_ONE)
         self.scenario_tree.add_child(self.behavior)
         self.scenario_tree.add_child(self.timeout_node)
         self.scenario_tree.add_child(self.criteria_tree)
@@ -107,11 +111,13 @@ class ScenarioManager(object):
     ego_vehicle = None
     other_actors = None
 
-    def __init__(self, world, _debug_mode):
+    def __init__(self, world, debug_mode=False):
         """
         Init requires scenario as input
         """
-        self._debug_mode = _debug_mode
+        self._debug_mode = debug_mode
+        self.agent = None
+        self._autonomous_agent_plugged = False
         self._running = False
         self._timestamp_last_run = 0.0
         self._my_lock = threading.Lock()
@@ -151,10 +157,11 @@ class ScenarioManager(object):
         self.end_system_time = None
         GameTime.restart()
 
-    def run_scenario(self):
+    def run_scenario(self, agent=None):
         """
         Trigger the start of the scenario and wait for it to finish/fail
         """
+        self.agent = agent
         print("ScenarioManager: Running scenario {}".format(self.scenario_tree.name))
         self.start_system_time = time.time()
         start_game_time = GameTime.get_time()
@@ -199,6 +206,11 @@ class ScenarioManager(object):
                 # Tick scenario
                 self.scenario_tree.tick_once()
 
+                if self.agent:
+                    # Invoke agent
+                    action = self.agent()
+                    self.ego_vehicle.apply_control(action)
+
                 if self._debug_mode:
                     print("\n")
                     py_trees.display.print_ascii_tree(
@@ -227,14 +239,21 @@ class ScenarioManager(object):
         failure = False
         timeout = False
         result = "SUCCESS"
-        for criterion in self.scenario.test_criteria:
-            if (not criterion.optional and
-                    criterion.test_status != "SUCCESS" and
-                    criterion.test_status != "ACCEPTABLE"):
+
+        if isinstance(self.scenario.test_criteria, py_trees.composites.Parallel):
+            if self.scenario.test_criteria.status == py_trees.common.Status.FAILURE:
                 failure = True
                 result = "FAILURE"
-            elif criterion.test_status == "ACCEPTABLE":
-                result = "ACCEPTABLE"
+        else:
+            for criterion in self.scenario.test_criteria:
+                if (not criterion.optional and
+                        criterion.test_status != "SUCCESS" and
+                        criterion.test_status != "ACCEPTABLE"):
+                    failure = True
+                    result = "FAILURE"
+                elif criterion.test_status == "ACCEPTABLE":
+                    result = "ACCEPTABLE"
+
 
         if self.scenario.timeout_node.timeout and not failure:
             timeout = True
@@ -244,3 +263,114 @@ class ScenarioManager(object):
         output.write()
 
         return failure or timeout
+
+    def analyze_scenario_challenge(self):
+        """
+        This function is intended to be called from outside and provide
+        statistics about the scenario (human-readable, for the CARLA challenge.)
+        """
+        PENALTY_COLLISION_STATIC = 10
+        PENALTY_COLLISION_VEHICLE = 10
+        PENALTY_COLLISION_PEDESTRIAN = 30
+        PENALTY_TRAFFIC_LIGHT = 10
+        PENALTY_WRONG_WAY = 5
+
+        target_reached = False
+        failure = False
+        result = "SUCCESS"
+        final_score = 0.0
+        score_penalty = 0.0
+        score_route = 0.0
+        return_message = ""
+
+        if isinstance(self.scenario.test_criteria, py_trees.composites.Parallel):
+            if self.scenario.test_criteria.status == py_trees.common.Status.FAILURE:
+                failure = True
+                result = "FAILURE"
+            if self.scenario.timeout_node.timeout and not failure:
+                result = "TIMEOUT"
+
+            list_traffic_events = []
+            for node in self.scenario.test_criteria.children:
+                if  node.list_traffic_events:
+                    list_traffic_events.extend(node.list_traffic_events)
+
+            list_collisions = []
+            list_red_lights = []
+            list_wrong_way = []
+            list_route_dev = []
+            # analyze all traffic events
+            for event in list_traffic_events:
+                if event.get_type() == TrafficEventType.COLLISION_STATIC:
+                    score_penalty += PENALTY_COLLISION_STATIC
+                    msg = event.get_message()
+                    if msg:
+                        list_collisions.append(event.get_message())
+
+                elif event.get_type() == TrafficEventType.COLLISION_VEHICLE:
+                    score_penalty += PENALTY_COLLISION_VEHICLE
+                    msg = event.get_message()
+                    if msg:
+                        list_collisions.append(event.get_message())
+
+                elif event.get_type() == TrafficEventType.COLLISION_PEDESTRIAN:
+                    score_penalty += PENALTY_COLLISION_PEDESTRIAN
+                    msg = event.get_message()
+                    if msg:
+                        list_collisions.append(event.get_message())
+
+                elif event.get_type() == TrafficEventType.TRAFFIC_LIGHT_INFRACTION:
+                    score_penalty += PENALTY_TRAFFIC_LIGHT
+                    msg = event.get_message()
+                    if msg:
+                        list_red_lights.append(event.get_message())
+
+                elif event.get_type() == TrafficEventType.WRONG_WAY_INFRACTION:
+                    score_penalty += PENALTY_WRONG_WAY
+                    msg = event.get_message()
+                    if msg:
+                        list_wrong_way.append(event.get_message())
+
+                elif event.get_type() == TrafficEventType.ROUTE_DEVIATION:
+                    msg = event.get_message()
+                    if msg:
+                        list_route_dev.append(event.get_message())
+
+                elif event.get_type() == TrafficEventType.ROUTE_COMPLETED:
+                    score_route = 100.0
+                    target_reached = True
+                elif event.get_type() == TrafficEventType.ROUTE_COMPLETION:
+                    if not target_reached:
+                        score_route = event.get_dict()['route_completed']
+
+            final_score = max(score_route - score_penalty, 0)
+
+            return_message += "\n=================================="
+            return_message += "\n==[{}] [Score = {:.2f} : (route_score={}, infractions=-{})]".format(result,
+                                                                                                   final_score,
+                                                                                                   score_route,
+                                                                                                   score_penalty)
+            if list_collisions:
+                return_message += "\n===== Collisions:"
+                for item in list_collisions:
+                    return_message += "\n========== {}".format(item)
+
+            if list_red_lights:
+                return_message += "\n===== Red lights:"
+                for item in list_red_lights:
+                    return_message += "\n========== {}".format(item)
+
+            if list_wrong_way:
+                return_message += "\n===== Wrong way:"
+                for item in list_wrong_way:
+                    return_message += "\n========== {}".format(item)
+
+            if list_route_dev:
+                return_message += "\n===== Route deviation:"
+                for item in list_route_dev:
+                    return_message += "\n========== {}".format(item)
+
+            return_message += "\n=================================="
+
+
+        return result, final_score, return_message