+++ srunner/utilities/code_check_and_formatting.sh
@@ -3,7 +3,10 @@
 autopep8 scenario_runner.py --in-place --max-line-length=120
 autopep8 srunner/scenariomanager/*.py --in-place --max-line-length=120
 autopep8 srunner/scenarios/*.py --in-place --max-line-length=120
+autopep8 srunner/challenge/*.py --in-place --max-line-length=120
+
 
 pylint --rcfile=.pylintrc --disable=I srunner/scenariomanager/
 pylint --rcfile=.pylintrc srunner/scenarios/
+pylint --rcfile=.pylintrc srunner/challenge/
 pylint --rcfile=.pylintrc scenario_runner.py